+++ src/cpu/x64/jit_avx512_core_amx_1x1_conv_kernel.cpp
@@ -259,7 +259,7 @@ void jit_avx512_core_amx_1x1_fwd_kernel_t::apply_postops(const Zmm zmm_out,
         }
 
         postops_injector_->compute_vector_range(
-                {(size_t)vmm_idx}, rhs_arg_params, ddp, qdp);
+                {vmm_idx}, rhs_arg_params, ddp, qdp);
 
         if ((jcp.with_depthwise || jcp.with_quantization)
                 && jcp.src_zero_point) {
